% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_versions.R
\name{consistent}
\alias{consistent}
\title{check consistency of package versions}
\usage{
consistent(R_version, packages, versions)
}
\arguments{
\item{R_version}{character vector of desired R version}

\item{packages}{character vector of desired package names}

\item{versions}{character vector of desired package versions}
}
\value{
A logical indicating consistence and completeness of submitted package versions.
}
\description{
check consistency of package versions
}
\examples{
    consistent('3.1.2', 'Rcpp', '1.0.2')
}
