% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{rockerfile}
\alias{rockerfile}
\title{create Dockerfile for desired R configuration}
\usage{
rockerfile(R_version, packages, versions)
}
\arguments{
\item{R_version}{character vector of desired R version}

\item{packages}{character vector of desired package names}

\item{versions}{character vector of desired package versions}
}
\value{
Dockerfile based on debian:stretch for desired R configuration
}
\description{
create Dockerfile for desired R configuration
}
\examples{
    rockerfile('3.1.2', 'Rcpp', '1.0.2')
}
